requirements.txt
discord.py
requests
python-dotenv
apscheduler
pytz
